requirements.txt
# Core dependencies
click>=8.1.0
pandas>=2.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
pyyaml>=6.0
dspy-ai>=2.0.0
openai>=1.0.0
nbformat>=5.9.0
toml>=0.10.2
scikit-learn>=1.3.0
matplotlib>=3.7.0
numpy>=1.24.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-timeout>=2.1.0

# Development dependencies
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
isort>=5.12.0

# Additional utilities
tqdm>=4.66.0
colorama>=0.4.6
python-dotenv>=1.0.0
